Remove finalizer
If we re-add the finalizer in the future, it will be simply to log a
warning in the case the secret was not closed prior to being garbage
collected.  But then we have to consider thread safety issues because
this class is not designed to be accessed from multiple threads, and the
finalizer thread is guaranteed to be different from whatever thread the
secret was instantiated upon.